+++ config/csvGen.js
@@ -5,8 +5,9 @@ const fs = require('fs');
 const path = require('path');
 
 // Winston Loggers
-const logger = require('./logs');
-const workerLog = logger.get('workerLog');
+const workerLog = require('./system/logs').get('workerLog');
+
+
 workerLog.info(`CSV ${threadId} Started`);
 
 // Write Stream Parameters
