+++ config/workerRelay.js
@@ -30,7 +30,7 @@ const inc = (n, poolName) => {
 
 
 // Winston Loggers
-const logger = require('./logs');
+const logger = require('./system/logs');
 const statsLog = logger.get('statsLog');
 
 // Initialize Porcessing Coverage Counter
